Rename panel to contents since it will have weights and items at some point.
